Rename AttendeeTable directory
